Renamed ScenarioSetup to TestDataBuilder and rename method names to Add*
